packages/database/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "database"
version = "0.1.0"
description = "Database models and migrations for Tiger MCP system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.28.0",
    "psycopg2-binary>=2.9.10",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "shared",
]

[tool.uv.sources]
shared = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/database"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
]

[dependency-groups]
dev = [
    "pytest-mock>=3.14.1",
    "pytest-cov>=6.2.1",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.28.1",
]